requirements.txt
streamlit==1.24.0
Pillow==9.5.0
opencv-python-headless==4.7.0
numpy==1.24.3
scikit-image==0.20.0
pandas==2.0.3
PyMuPDF==1.22.3
python-docx==0.8.11
requests==2.31.0
